+++ src/DXMemory.c
@@ -53,34 +53,34 @@ void* dx_error_processor (void* src) {
  */
 /* -------------------------------------------------------------------------- */
 
-void* dx_malloc (int size) {
+void* dx_malloc (size_t size) {
 	void* r = malloc(size);
     return dx_error_processor(r);
 }
 
 /* -------------------------------------------------------------------------- */
 
-void* dx_calloc (int num, int size) {
+void* dx_calloc (size_t num, size_t size) {
 	void* r = calloc(num, size);
     return dx_error_processor(r);
 }
 
 /* -------------------------------------------------------------------------- */
 
-void* dx_memcpy (void * destination, const void * source, int size) {
+void* dx_memcpy (void * destination, const void * source, size_t size) {
     
     return dx_error_processor(memcpy(destination, source, size));
 }
 
 /* -------------------------------------------------------------------------- */
 
-void* dx_memmove (void * destination, const void * source, int size) {
+void* dx_memmove (void * destination, const void * source, size_t size) {
     return dx_error_processor(memmove(destination, source, size));
 }
 
 /* -------------------------------------------------------------------------- */
 
-void* dx_memset (void * destination, int c, int size) {
+void* dx_memset (void * destination, int c, size_t size) {
     return dx_error_processor(memset(destination, c, size));
 }
 
@@ -96,7 +96,7 @@ void dx_free (void* buf) {
  */
 /* -------------------------------------------------------------------------- */
 
-void* dx_calloc_no_ehm (int num, int size) {
+void* dx_calloc_no_ehm (size_t num, size_t size) {
 	void *r = calloc(num, size);
     return r;
 }
